requirements.txt
Django>=4.2
ultralytics
gunicorn
Pillow
torch
